mute [nfc]: Move shouldBeMuted to where it's used.
This function is effectively really a private helper for its caller.
