+++ src/chat/narrowsSelectors.js
@@ -6,6 +6,7 @@ import { createSelector } from 'reselect';
 import type {
   GlobalState,
   Message,
+  MuteState,
   Narrow,
   Outbox,
   Selector,
@@ -24,12 +25,16 @@ import { getCaughtUpForNarrow } from '../caughtup/caughtUpSelectors';
 import { getAllUsersById, getOwnUserId } from '../users/userSelectors';
 import {
   isStreamOrTopicNarrow,
+  isHomeNarrow,
+  isTopicNarrow,
+  isMentionedNarrow,
   isMessageInNarrow,
   caseNarrowDefault,
   keyFromNarrow,
   streamNameOfNarrow,
 } from '../utils/narrow';
-import { shouldBeMuted } from '../utils/message';
+import { isTopicMuted } from '../mute/muteModel';
+import { streamNameOfStreamMessage } from '../utils/recipient';
 import { NULL_ARRAY, NULL_SUBSCRIPTION } from '../nullObjects';
 import * as logging from '../utils/logging';
 
@@ -91,6 +96,49 @@ export const getMessagesForNarrow: Selector<$ReadOnlyArray<Message | Outbox>, Na
       return [...fetchedMessages, ...outboxMessages].sort((a, b) => a.id - b.id);
     },
   );
+
+/** Private helper; exported only for tests. */
+export const shouldBeMuted = (
+  message: Message | Outbox,
+  narrow: Narrow,
+  subscriptions: $ReadOnlyArray<Subscription>,
+  mutes: MuteState,
+): boolean => {
+  if (message.type === 'private') {
+    return false; // private/group messages are not muted
+  }
+
+  if (isTopicNarrow(narrow)) {
+    return false; // never hide a message when narrowed to topic
+  }
+
+  // This logic isn't quite right - we want to make sure we never hide a
+  // message that has a mention, even if we aren't in the "Mentioned"
+  // narrow. (#3472)  However, it's more complex to do that, and this code
+  // fixes the largest problem we'd had with muted mentioned messages, which
+  // is that they show up in the count for the "Mentions" tab, but without
+  // this conditional they wouldn't in the actual narrow.
+  if (isMentionedNarrow(narrow)) {
+    return false;
+  }
+
+  const streamName = streamNameOfStreamMessage(message);
+
+  if (isHomeNarrow(narrow)) {
+    const sub = subscriptions.find(x => x.name === streamName);
+    if (!sub) {
+      // If there's no matching subscription, then the user must have
+      // unsubscribed from the stream since the message was received.  Leave
+      // those messages out of this view, just like for a muted stream.
+      return true;
+    }
+    if (!sub.in_home_view) {
+      return true;
+    }
+  }
+
+  return isTopicMuted(streamName, message.subject, mutes);
+};
 
 // Prettier mishandles this Flow syntax.
 // prettier-ignore
